change directory semi-functional. cannot change into test_folder though without absolute path, its as if the parent process is running in a different directory

+++ assign2_25201495/helpers.h
@@ -1,10 +1,11 @@
 #ifndef HELPERS_H
 #define HELPERS_H
+#include <stddef.h>
 
 void display_prompt();
 struct tm* get_current_time();
 char** parse_and_store_tokens(char *line);
 void execute_chdir(char **args);
-static int count_strings(char **arr);
+static size_t count_strings(char **arr);
 
 #endif
